requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
factory_boy==3.3.3
Faker==36.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tenacity==9.0.0
tzdata==2025.1
urllib3==2.3.0
